django/requirements.txt
cffi==1.15.1
cryptography==41.0.7
Django==2.1.15
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
ecdsa==0.18.0
psycopg2-binary==2.8.5
pyasn1==0.5.1
pycparser==2.21
PyJWT==1.7.1
python-dotenv==0.21.1
python-jose==3.3.0
pytz==2023.3.post1
rsa==4.9
six==1.16.0
uWSGI==2.0.20
